Rigging in logical expression Cursors into CursorBuilder ...
 o fixed bug in optimization code for Evaluators where Long scan counts should 
   be used instead of Integer scan counts
 o fixed generics problem in SubstringCursor
 o using Lists instead of arrays for child Cursors and Evaluators in the 
   OrCursor - again this is due to problems with using generics to create
   generic arrays
 o added code to build following Cursors in CursorBuilder which is still badly
   broken and will not compile:
     - OrCursor
     - AndCursor
     - NotCursor
     - SubstringCursor



git-svn-id: https://svn.apache.org/repos/asf/directory/sandbox/akarasulu/bigbang/apacheds@642511 13f79535-47bb-0310-9956-ffa450edef68

+++ xdbm-search/src/main/java/org/apache/directory/server/xdbm/search/impl/AndEvaluator.java
@@ -72,8 +72,8 @@ public class AndEvaluator implements Evaluator<AndNode,Attributes>
         {
             public int compare( Evaluator<?, Attributes> e1, Evaluator<?, Attributes> e2 )
             {
-                int scanCount1 = ( Integer ) e1.getExpression().get( "count" );
-                int scanCount2 = ( Integer ) e2.getExpression().get( "count" );
+                long scanCount1 = ( Long ) e1.getExpression().get( "count" );
+                long scanCount2 = ( Long ) e2.getExpression().get( "count" );
 
                 if ( scanCount1 == scanCount2 )
                 {
